Copy to and from demand memory buffer.
Virt doesn't have the tooling for internal copies, or copies to and from
a write buffer that might exist outside of virt's reserved memory.

With more complicated FTLs that might evict and later read pages to disk
from a memory buffer, or serve requests inside the memory buffer, we
need some way to do that without writing the data to virt's disk memory.

Basic callback functionality is also added to internal work, which was originally
used to support DFTL KV's asynchronous processing.