Initial switch to workers.
Before, a lot of work was being done in the single dispatcher thread.
This begins the switch to giving the workers all of the work.

In general, anything that isn't just getting a timing should be
in a worker thread, not the dispatcher.

The demand based scheme has to do it a little differently than the conventional
scheme, though. The skiplist needs the write data for an insert,
but we don't want to do that work in the dispatcher thread.
If we just advance the PCIe timestamp in the dispatcher thread, and
place the skiplist insert on a background worker to execute before the
write completes to the user, then it doesn't seem there's any guarantee
the write will have copied to the skiplist by then.

We can just place the skiplist on the background thread first, set
the write's completion timestamp to far in the future, give the original
request to the skiplist for later, and change the completion time of
that request when the insert finishes. Note the skiplist insert might
turn into a full write buffer flush too.