Proper work handling in DFTLKV.
The goal is to have a work item in the queue representing the user's
request (as normal in NVMeV), which we use for the completion, and have
other background work doing the actual KV functions. For this, we need
a work item that's already in the queue that we can set the target time
for when a KV function is complete. This commit changes things up when
using DFTLKV to place the work item in the queue first, with a very high
timestamp. That work item is given to the background work item (doing
the actual KV work), then its timestamp is set to the actual completion
timestamp when the KV work is done.