Place a spinlock on write buffer transfers.
In DFTLKV, timings are generated in the worker threads, not the single
dispatcher.

When multiple worker threads on different shards need to write,
they share the same PCIe channel. Each write will get a start time
in demand_write, then use that start time as the start time for the
PCIe channel request.

If we don't lock here, different threads will access the same PCIe
channel
structure at the same time, which isn't supported right now.

However, we can't place the spin in ssd_advance_pcie because of the
following situation:

Thread 1 shard 0 gets start time 0 in demand_write
Thread 2 shard 1 gets start time 1 in demand_write

Thread 2 locks the spin in ssd_advance_pcie, and models the request.
This mean the last request time on the PCIe channel is 1.

Thread 2 unlocks the spin, and thread 1 gets it. It tries to advance
the PCIe channel with a start time of 0, which is less than the
start time of the previous request. The current channel model assumes
requests arrive with a start time equal to or greater than the time
of the
previous request.

Therefore, we use a spinlock here to ensure that the start times
for write buffer transfers are in order.