Initial work into fast, single-thread design
This commit introduces the initial work into making a fast,
single-threaded design, as opposed to one where all of the work is done
on background workers. This design follows the original NVMeVirt
conventional FTL design where all of the indexing work is done in the
dispatcher thread, with copies being done in the background.

The motivation is this; if we can re-write the FTL to be fast
enough with a single thread, it greatly simplifies its design and
extension. Here, "fast enough" means with very low flash
timings we can get many more IOPS than we would see in real life. Then,
we can configure appropriate flash timings to get realistic performance.

Moreover, using the single-thread design means we don't have to use
hacks around the code, such as placing fake work requests on the
background threads, and updating them later. The code was getting
confusing, and the amount of ifdefs for DFTLKV excessive.

There is one outstanding question with this design, and it's the
question of if the KVSSD FTL overhead (e.g. hashing) is appropriately
modeled. If the CPU/memory frequencies of the dispatcher thread are
lowered in accordance with the target device, then I don't see why not
at the moment.