Working GC.
The major difference is that we don't invalidate mapping pages as soon
as they are read into memory, like in the original implementation.

We don't invalidate these pages because doing so would mean the area of
NVMeVirt's reserved memory we use for the pages could be re-used for
another user data or mapping page during GC. We favor using NVMeVirt's
reserved memory to avoid allocations.