Initial work on the new caching scheme.
New caching scheme where only the written LPA to PPA mappings are stored.
Really, the original caching scheme (each map entry is a full
page) can also be implemented as a version of this scheme that just
tries to evict and read a full page each time, instead of only the
used grains. After the new scheme is worked out we can do that too.