Use the Kernel's allocator instead of Virt's mem.
It isn't actually necessary for us to tie a KV pair's placement in
memory to the PPA its assigned to. The PPAs are there to generate flash
timings. When we tie the location of KV pairs to their PPA, it means
during garbage collection we have to copy everything to a new location,
putting a ton of stress on the system and increasing complexity.
If we just allocate KV pair data with the kernel allocator, we never
need to move it unless the size changes. Originally, it's likely
NVMeVirt allocated memory up-front so that it could be managed manually
for performance reasons, but this change actually increases performance
already, compared to how we did it previously.

Why do it this way for the KV version and not the block version? In the
block version, data is copied to NVMeVirt's memory + LPA offset always.
That is really simple to manage, because we never need to resize and
copy data elsewhere. We could write KV pairs to a location in NVMeVirt's
reserved memory, and only move them when the size changes, but then
we're just rewriting the allocator for no good reason (yet).