Beginnings of append with big values.
Keeps a single append buffer in memory that is only flushed when a new
append buffer for a new key is created. Currently appends to the end of
the buffer, but we will probably have to change this to be grain-based
so that we can delete grains in the larger values using offset delete.